Get rid of ludicrous mode
