+++ dgraph/cmd/alpha/run.go
@@ -211,14 +211,6 @@ they form a Raft group and provide synchronous replication.
 			"Number of maximum pending queries before we reject them as too many requests.").
 		String())
 
-	flag.String("ludicrous", worker.LudicrousDefaults, z.NewSuperFlagHelp(worker.LudicrousDefaults).
-		Head("Ludicrous options").
-		Flag("enabled",
-			"Set enabled to true to run Dgraph in Ludicrous mode.").
-		Flag("concurrency",
-			"The number of concurrent threads to use in Ludicrous mode.").
-		String())
-
 	flag.String("graphql", worker.GraphQLDefaults, z.NewSuperFlagHelp(worker.GraphQLDefaults).
 		Head("GraphQL options").
 		Flag("introspection",
@@ -695,8 +687,6 @@ func run() {
 	tlsServerConf, err := x.LoadServerTLSConfigForInternalPort(Alpha.Conf)
 	x.Check(err)
 
-	ludicrous := z.NewSuperFlag(Alpha.Conf.GetString("ludicrous")).MergeAndCheckDefault(
-		worker.LudicrousDefaults)
 	raft := z.NewSuperFlag(Alpha.Conf.GetString("raft")).MergeAndCheckDefault(worker.RaftDefaults)
 	x.WorkerConfig = x.WorkerOptions{
 		TmpDir:              Alpha.Conf.GetString("tmp"),
@@ -708,8 +698,6 @@ func run() {
 		AclEnabled:          aclKey != nil,
 		AbortOlderThan:      abortDur,
 		StartTime:           startTime,
-		Ludicrous:           ludicrous,
-		LudicrousEnabled:    ludicrous.GetBool("enabled"),
 		Security:            security,
 		TLSClientConfig:     tlsClientConf,
 		TLSServerConfig:     tlsServerConf,
