+++ dgraph/cmd/live/batch.go
@@ -257,7 +257,7 @@ func (l *loader) conflictKeysForNQuad(nq *api.NQuad) ([]uint64, error) {
 	pred, found := l.schema.preds[attr]
 
 	// We dont' need to generate conflict keys for predicate with noconflict directive.
-	if found && pred.NoConflict || opt.ludicrousMode {
+	if found && pred.NoConflict {
 		return nil, nil
 	}
