+++ systest/21million/live/run_test.go
@@ -18,6 +18,9 @@ package bulk
 
 import (
 	"os"
+	"path/filepath"
+
+	"github.com/dgraph-io/dgraph/testutil"
 
 	"github.com/dgraph-io/dgraph/systest/21million/common"
 
@@ -28,21 +31,21 @@ func TestQueries(t *testing.T) {
 	t.Run("Run queries", common.TestQueriesFor21Million)
 }
 
-// func TestMain(m *testing.M) {
-// 	schemaFile := filepath.Join(testutil.TestDataDirectory, "21million.schema")
-// 	rdfFile := filepath.Join(testutil.TestDataDirectory, "21million.rdf.gz")
-// 	if err := testutil.LiveLoad(testutil.LiveOpts{
-// 		Alpha:      testutil.ContainerAddr("alpha1", 9080),
-// 		Zero:       testutil.SockAddrZero,
-// 		RdfFile:    rdfFile,
-// 		SchemaFile: schemaFile,
-// 	}); err != nil {
-// 		cleanupAndExit(1)
-// 	}
+func TestMain(m *testing.M) {
+	schemaFile := filepath.Join(testutil.TestDataDirectory, "21million.schema")
+	rdfFile := filepath.Join(testutil.TestDataDirectory, "21million.rdf.gz")
+	if err := testutil.LiveLoad(testutil.LiveOpts{
+		Alpha:      testutil.ContainerAddr("alpha1", 9080),
+		Zero:       testutil.SockAddrZero,
+		RdfFile:    rdfFile,
+		SchemaFile: schemaFile,
+	}); err != nil {
+		cleanupAndExit(1)
+	}
 
-// 	exitCode := m.Run()
-// 	cleanupAndExit(exitCode)
-// }
+	exitCode := m.Run()
+	cleanupAndExit(exitCode)
+}
 
 func cleanupAndExit(exitCode int) {
 	_ = os.RemoveAll("./t")
