--- systest/21million/ludicrous/docker-compose.yml
@@ -1,35 +1,0 @@
-# Auto-generated with: [./compose -a 3 -z 1 -r 1 -d data -w]
-#
-version: "3.5"
-services:
-  alpha1:
-    image: dgraph/dgraph:latest
-    working_dir: /data/alpha1
-    labels:
-      cluster: test
-    ports:
-    - 8080
-    - 9080
-    volumes:
-    - type: bind
-      source: $GOPATH/bin
-      target: /gobin
-      read_only: true
-    command: /gobin/dgraph alpha --my=alpha1:7080 --zero=zero1:5080
-      --logtostderr -v=2 --raft="idx=1;" 
-      --security "whitelist=10.0.0.0/8,172.16.0.0/12,192.168.0.0/16;"
-      --ludicrous "enabled=true;"
-  zero1:
-    image: dgraph/dgraph:latest
-    working_dir: /data/zero1
-    labels:
-      cluster: test
-    ports:
-    - 5080
-    - 6080
-    volumes:
-    - type: bind
-      source: $GOPATH/bin
-      target: /gobin
-      read_only: true
-    command: /gobin/dgraph zero --raft="idx=1;" --my=zero1:5080 --logtostderr -v=2 --bindall
