--- systest/ludicrous/1predicate-c.schema
@@ -1,1 +1,0 @@
-pred: [int] @count .
