--- systest/ludicrous/1predicate.schema
@@ -1,1 +1,0 @@
-pred: [int] .
