--- systest/ludicrous/docker-compose.yml
@@ -1,81 +1,0 @@
-# Auto-generated with: [./compose -l -a 3 -z 1 -r 1 --alpha_volume data:/data --port_offset=0 --expose_ports=false --mem= --names=false -O ../systest/ludicrous/docker-compose.yml]
-#
-version: "3.5"
-services:
-  alpha1:
-    image: dgraph/dgraph:latest
-    working_dir: /data/alpha1
-    labels:
-      cluster: test
-    ports:
-    - "8080"
-    - "9080"
-    volumes:
-    - type: bind
-      source: $GOPATH/bin
-      target: /gobin
-      read_only: true
-    - type: volume
-      source: data
-      target: /data
-      read_only: false
-    command: /gobin/dgraph alpha --my=alpha1:7080 --zero=zero1:5080 --logtostderr
-      -v=2 
-      --security "whitelist=10.0.0.0/8,172.16.0.0/12,192.168.0.0/16;"
-  alpha2:
-    image: dgraph/dgraph:latest
-    working_dir: /data/alpha2
-    labels:
-      cluster: test
-    ports:
-    - "8080"
-    - "9080"
-    volumes:
-    - type: bind
-      source: $GOPATH/bin
-      target: /gobin
-      read_only: true
-    - type: volume
-      source: data
-      target: /data
-      read_only: false
-    command: /gobin/dgraph alpha --my=alpha2:7080 --zero=zero1:5080 --logtostderr
-      -v=2 
-      --security "whitelist=10.0.0.0/8,172.16.0.0/12,192.168.0.0/16;"
-  alpha3:
-    image: dgraph/dgraph:latest
-    working_dir: /data/alpha3
-    labels:
-      cluster: test
-    ports:
-    - "8080"
-    - "9080"
-    volumes:
-    - type: bind
-      source: $GOPATH/bin
-      target: /gobin
-      read_only: true
-    - type: volume
-      source: data
-      target: /data
-      read_only: false
-    command: /gobin/dgraph alpha --my=alpha3:7080 --zero=zero1:5080 --logtostderr
-      -v=2 
-      --security "whitelist=10.0.0.0/8,172.16.0.0/12,192.168.0.0/16;"
-  zero1:
-    image: dgraph/dgraph:latest
-    working_dir: /data/zero1
-    labels:
-      cluster: test
-    ports:
-    - "5080"
-    - "6080"
-    volumes:
-    - type: bind
-      source: $GOPATH/bin
-      target: /gobin
-      read_only: true
-    command: /gobin/dgraph zero --raft="idx=1" --my=zero1:5080 --replicas=1 --logtostderr
-      -v=2 --bindall
-volumes:
-  data: {}
