--- systest/ludicrous/queries/query-001
@@ -1,13 +1,0 @@
-{
-  result(func: gt(count(pred), 5)) {
-    count(pred) 
-  }
-}    
----
-{
-  "result": [
-    {
-      "count(pred)": 30000
-    }
-  ]
-}
