+++ worker/draft.go
@@ -75,8 +75,6 @@ type node struct {
 	canCampaign bool
 	elog        trace.EventLog
 
-	ex *executor
-
 	keysWritten      *keysWritten
 	pendingProposals []pb.Proposal
 }
@@ -297,9 +295,6 @@ func newNode(store *raftwal.DiskStorage, gid uint32, id uint64, myAddr string) *
 		cdcTracker:  newCDC(),
 		keysWritten: newKeysWritten(),
 	}
-	if x.WorkerConfig.LudicrousEnabled {
-		n.ex = newExecutor(&m.Applied, int(x.WorkerConfig.Ludicrous.GetInt64("concurrency")))
-	}
 	return n
 }
 
@@ -390,11 +385,6 @@ func (n *node) processMutations(ctx context.Context, m *pb.Mutations) {
 		}
 		return ei.GetEntity() < ej.GetEntity()
 	})
-
-	// if x.WorkerConfig.LudicrousEnabled {
-	// 	n.ex.addEdges(ctx, proposal)
-	// 	return nil
-	// }
 
 	span := otrace.FromContext(ctx)
 	txn := posting.Oracle().GetTxn(m.StartTs)
@@ -543,16 +533,6 @@ func (n *node) applyMutations(ctx context.Context, proposal *pb.Proposal) (rerr 
 			if err := detectPendingTxns(supdate.Predicate); err != nil {
 				return err
 			}
-		}
-
-		// If Dgraph is running in ludicrous mode and we get some schema we should wait for all
-		// active mutations to finish. Previously we were thinking of only waiting for active
-		// mutations related to predicates present in schema mutation. But this might cause issues
-		// as we call DropPrefix() on Badger while running schema mutations. DropPrefix() blocks
-		// writes on Badger and returns error if writes are tried. To avoid this we should wait for
-		// all active mutations to finish irrespective of predicates present in schema mutation.
-		if x.WorkerConfig.LudicrousEnabled && len(proposal.Mutations.Schema) > 0 {
-			n.ex.waitForActiveMutations()
 		}
 
 		if err := runSchemaMutation(ctx, proposal.Mutations.Schema, startTs); err != nil {
@@ -817,15 +797,6 @@ func (n *node) processApplyCh() {
 
 	// This function must be run serially.
 	handle := func(prop pb.Proposal) {
-		// Ignore the start ts in case of ludicrous mode. We get a new ts and use that as the
-		// commit ts.
-		// WARNING: This would cause the leader and the follower to diverge in the timestamp
-		// they use to commit the same thing.
-		// TODO: This is broken. We need to find a way to fix this.
-		if x.WorkerConfig.LudicrousEnabled && prop.Mutations != nil {
-			prop.Mutations.StartTs = State.GetTimestamp(false)
-		}
-
 		var perr error
 		prev, ok := previous[prop.Key]
 		if ok && prev.err == nil {
@@ -972,14 +943,8 @@ func (n *node) commitOrAbort(pkey uint64, delta *pb.OracleDelta) error {
 	for _, status := range delta.Txns {
 		toDisk(status.StartTs, status.CommitTs)
 	}
-	if x.WorkerConfig.LudicrousEnabled {
-		if err := writer.Wait(); err != nil {
-			glog.Errorf("Error while waiting to commit: +%v", err)
-		}
-	} else {
-		if err := writer.Flush(); err != nil {
-			return errors.Wrapf(err, "while flushing to disk")
-		}
+	if err := writer.Flush(); err != nil {
+		return errors.Wrapf(err, "while flushing to disk")
 	}
 	if x.WorkerConfig.HardSync {
 		if err := pstore.Sync(); err != nil {
@@ -1244,9 +1209,6 @@ func (n *node) checkpointAndClose(done chan struct{}) {
 				time.Sleep(time.Second) // Let transfer happen.
 			}
 			n.Raft().Stop()
-			if x.WorkerConfig.LudicrousEnabled {
-				n.ex.closer.SignalAndWait()
-			}
 			close(done)
 			return
 		}
@@ -1482,22 +1444,6 @@ func (n *node) Run() {
 						atomic.AddUint32(&pctx.Found, 1)
 						if span := otrace.FromContext(pctx.Ctx); span != nil {
 							span.Annotate(nil, "Proposal found in CommittedEntries")
-						}
-						if x.WorkerConfig.LudicrousEnabled {
-							var p pb.Proposal
-							if err := p.Unmarshal(entry.Data[8:]); err != nil {
-								glog.Errorf("Unable to unmarshal proposal: %v %x\n",
-									err, entry.Data)
-								break
-							}
-							if len(p.Mutations.GetEdges()) > 0 {
-								// Assuming that there will be no error while applying. But this
-								// assumption is only made for data mutations and not schema
-								// mutations.
-								// TODO: This should not be done here. Instead, it should be done
-								// within the ludicrous mode scheduler.
-								n.Proposals.Done(key, nil)
-							}
 						}
 					}
 					entries = append(entries, entry)
@@ -1838,10 +1784,7 @@ func (n *node) calculateSnapshot(startIdx, lastIdx, minPendingStart uint64) (*pb
 					snapshotIdx = entry.Index - 1
 				}
 			}
-			// In ludicrous mode commitTs for any transaction is same as startTs.
-			if x.WorkerConfig.LudicrousEnabled {
-				maxCommitTs = x.Max(maxCommitTs, start)
-			} else if proposal.Delta != nil {
+			if proposal.Delta != nil {
 				for _, txn := range proposal.Delta.GetTxns() {
 					maxCommitTs = x.Max(maxCommitTs, txn.CommitTs)
 				}
