+++ worker/server_state.go
@@ -38,13 +38,12 @@ const (
 	//       For easy readability, keep the options without default values (if any) at the end of
 	//       the *Defaults string. Also, since these strings are printed in --help text, avoid line
 	//       breaks.
-	AclDefaults       = `access-ttl=6h; refresh-ttl=30d; secret-file=;`
-	AuditDefaults     = `compress=false; days=10; size=100; dir=; output=; encrypt-file=;`
-	BadgerDefaults    = `compression=snappy; goroutines=8; max-retries=-1;`
-	RaftDefaults      = `learner=false; snapshot-after=10000; pending-proposals=256; idx=; group=;`
-	SecurityDefaults  = `token=; whitelist=;`
-	LudicrousDefaults = `enabled=false; concurrency=2000;`
-	CDCDefaults       = `file=; kafka=; sasl_user=; sasl_password=; ca_cert=; client_cert=; ` +
+	AclDefaults      = `access-ttl=6h; refresh-ttl=30d; secret-file=;`
+	AuditDefaults    = `compress=false; days=10; size=100; dir=; output=; encrypt-file=;`
+	BadgerDefaults   = `compression=snappy; goroutines=8; max-retries=-1;`
+	RaftDefaults     = `learner=false; snapshot-after=10000; pending-proposals=256; idx=; group=;`
+	SecurityDefaults = `token=; whitelist=;`
+	CDCDefaults      = `file=; kafka=; sasl_user=; sasl_password=; ca_cert=; client_cert=; ` +
 		`client_key=;`
 	LimitDefaults = `mutations=allow; query-edge=1000000; normalize-node=10000; ` +
 		`mutations-nquad=1000000; disallow-drop=false; max-pending-queries=10000; query-timeout
